fix(api): unit tests.
