+++ api/api/tests/vuln_tests.py
@@ -3,9 +3,6 @@ from django.test import TransactionTestCase
 from rest_framework.test import APIClient
 from api.models import Manager, Pentester
 
-from api.management.commands.init_builtin_vuln_types import Command
-from api.models import Manager, Pentester
-
 from api.tests.helpers import create_mission, create_random_pentester, create_random_manager, default_user_password, \
     login_as
